*Separated types logic into separate files for better organization and maintainability
